Add files via upload
Iterations and Fractals is the second part in a series of four reports that explores the implementation of Newton’s Method to find the roots of functions and create fractals in the complex plane. Additionally, this laboratory examines the convergence of the solutions of the algorithm, comparing the expected and experimental errors.